decorater分離
